+++ NEWS.md
@@ -1,6 +1,6 @@
 # raadtools dev
 
-* First parts of REMA data now available from `readtopo`, with "rema_100m" and `read_rema_tiles` for the 8m index. 
+* First parts of REMA data now available from `readtopo`, with "rema_8m", "rema_100m" and `read_rema_tiles` (the 8m index). 
 
 * AMPS data now supports basic access for 10km "d2" files, with `amps_d1files` moved to raadfiles.